requirements.txt
flask
openai
python-dotenv
gunicorn
threading
requests